requirements.txt
streamlit
google-generativeai
PyPDF2
faiss-cpu
python-dotenv
tqdm
sentence-transformers
faiss-cpu
pandas
matplotlib
seaborn
numpy

